added permissions

+++ public/js/custom/getProducts.js
@@ -44,7 +44,7 @@ $(function () {
     // Advanced Search
     if (dt_adv_filter_table.length) {
         var dt_adv_filter = dt_adv_filter_table.DataTable({
-            ajax: `${window.location.origin}/survey/public/getproducts`,
+            ajax: `${baseurl}/getproducts`,
             order: [[5, "desc"]],
             columns: [
                 {
@@ -68,8 +68,8 @@ $(function () {
                     render: function (value) {
                         if (value === null) return "";
                         return `<span class="badge rounded-pill badge-light-${
-                            value == 0 ? "danger" : "success"
-                        }"}>${value == 0 ? "Inactive" : "Active"}</span>`;
+                            value === 0 ? "danger" : "success"
+                        }"}>${value === 0 ? "Inactive" : "Active"}</span>`;
                     },
                 },
                 {
@@ -91,6 +91,7 @@ $(function () {
                     "orderable": false,
                     render: function (value) {
                         if (value === null) return "";
+                        if (!showDeleteIcon) return "";
                         return `<div class="d-flex flex-wrap align-items-center">
                                     <span onclick="deleteProduct(${value})" class="cursor-pointer">
                                         ${feather.icons["trash"].toSvg({
@@ -147,5 +148,31 @@ function confirmDeleteProduct() {
     let deleteProductForm = $("#deleteProductForm");
     if (deleteProductForm.length) {
         deleteProductForm.submit();
+    }
+}
+
+function getEditIcon(iconType, compId = 0) {
+    // 1 => show edit icon
+    // 2 => show delete icon
+    switch (iconType) {
+        case 1:
+            return showEditIcon
+                ? `<a href="${baseurl}/editcompany/${compId}">
+        ${feather.icons["edit"].toSvg({
+            class: "me-1",
+        })}
+    </a>`
+                : "";
+        case 2:
+            return showDeleteIcon
+                ? ` <span onclick="deleteCompany(${compId})" class="cursor-pointer">
+                ${feather.icons["trash"].toSvg({
+                    class: "text-primary",
+                })}
+            </span>`
+                : "";
+
+        default:
+            break;
     }
 }
